# optimize saveVector and restoreVector (#10617)