rewrite 
 writeStringViews and restoreVectorStringViews 


before the 1000k can't run:


[...]s/benchmarks/VectorSaverBenchmark.cpp     relative  time/iter   iters/s
save_vector_10k                                            58.76ms     17.02
save_vector_100k                                             7.96s   125.67m
restore_vector_10k                                         22.15ms     45.15
restore_vector_100k                                          4.60s   217.40m


`
after:



[...]s/benchmarks/VectorSaverBenchmark.cpp     relative  time/iter   iters/s
save_vector_10k                                           170.42us     5.87K
save_vector_100k                                            2.51ms    399.06
save_vector_1000k                                          39.88ms     25.08
restore_vector_10k                                         25.00us    40.00K
restore_vector_100k                                       274.83us     3.64K
restore_vector_10000k                                       4.98ms    200.79
`